Fix ConsistencyModel int casts and add search space TODOs
- consistency.py: Fix s0, s1, max_time casts from float() to int()
  to match their IntDimension declarations and BayesFlow's expected
  parameter types (TODO #27)
- registry.py: Add return type annotations to get_inference_space()
  and get_summary_space()
- docs/TODO.md: Add search space TODOs 27-29 (int cast bug now
  resolved, IntDimension log+step validation, FusionTransformer
  missing dimensions)

+++ src/bayesflow_hpo/search_spaces/inference/consistency.py
@@ -121,12 +121,12 @@ class ConsistencyModelSpace(BaseSearchSpace):
             },
         }
         if "cm_max_time" in params:
-            kwargs["max_time"] = float(params["cm_max_time"])
+            kwargs["max_time"] = int(params["cm_max_time"])
         if "cm_sigma2" in params:
             kwargs["sigma2"] = float(params["cm_sigma2"])
         if "cm_s0" in params:
-            kwargs["s0"] = float(params["cm_s0"])
+            kwargs["s0"] = int(params["cm_s0"])
         if "cm_s1" in params:
-            kwargs["s1"] = float(params["cm_s1"])
+            kwargs["s1"] = int(params["cm_s1"])
 
         return bf.networks.ConsistencyModel(**kwargs)
